requirements.txt
colorlog==6.8.0
dj_database_url==2.1.0
Django==5.0
django-import-export==3.3.3
django-model-utils==4.3.1
django-split-settings==1.2.0
django-storages==1.14.2
django-stubs==4.2.7
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.1
drf-spectacular==0.26.5
gunicorn==21.2.0
psycopg[c]==3.1.14
pyjwt==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
